pyproject.toml
[project]
name = "skiing-watermark-remover"
version = "0.3.0"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "torch>=2.0.0",
    "diffusers>=0.25.0",
    "transformers>=4.36.0", # GroundingDINO support
    "accelerate>=0.20.0",
    "segment-anything>=1.0",
    "timm>=0.9.0", # Required for GroundingDINO
    "addict>=2.4.0", # Often needed for config parsing
    "yapf>=0.40.0", # Often needed
    "aiohttp>=3.13.2",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true
